refactor(loops): lift data-row media resolution into its own module
Merging upstream put this branch's filter/order plumbing and #300's wider
media resolution into the same file, and the result crossed the 700-line
module ceiling at 711 — CI caught it.

Raising the ceiling would have been the wrong repair. The two halves
answer different questions: the rest of the file decides WHICH rows a
loop returns — the filter, the order, the page window — while this block
decides what a row's media cells CONTAIN once those rows are in hand.
They only ever met because both needed the same page of rows.

So `resolveMediaIdsToPaths`, `collectMediaIds` and `resolvedMediaOverlay`
move to `dataRowsMedia.ts` unchanged, and the source file imports them.
No behaviour changes: 596 lines and 132 rather than 711.

`MediaAssetRow` and the `readMediaCellIds` / `readRepeaterCell` imports
went with them, and the one test that reached for `resolveMediaIdsToPaths`
now imports it from where it lives.
